Only report reasons sending remaining and IAP issue. else all done for version 2

+++ Kismat App/Views/ViewControllers/SocialLinksView/SocialLinks VC.swift
@@ -6,20 +6,34 @@
 //
 
 import UIKit
+import UIMultiPicker
 
 class SocialLinks_VC: MainViewController {
 
     
     @IBOutlet weak var headingLbl: fullyCustomLbl!
     @IBOutlet weak var socialLinksTV: UITableView!
+    @IBOutlet weak var multiPickerView: UIMultiPicker!
+    @IBOutlet weak var pickerView: UIView!
+    
+    @IBAction func doneBtnPressed(_ sender: Any) {
+        pickerView.isHidden = true
+    }
+    
     
     var socialAccModel = [SocialAccModel]()
     var linkType = ""
     var canEdit = false
-    
+    var isFromOther = false
+    var reasonsList = [ReportReasonsModel]()
+    var reasonsListName = [String]()
+    var selectedReasonsAray = [Int]()
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        getReasons()
+        
         headingLbl.text = "Social Links Connected"
         registerCells()
     }
@@ -39,12 +53,68 @@ class SocialLinks_VC: MainViewController {
 
     }
 
+    func setupMultiPickerView() {
+        
+        pickerView.isHidden = false
+        multiPickerView.options = reasonsListName
+        
+        multiPickerView.addTarget(self, action: #selector(selected(_:)), for: .valueChanged)
+        
+        multiPickerView.color = .darkGray
+        multiPickerView.tintColor = .black
+        multiPickerView.font = .systemFont(ofSize: 18, weight: .semibold)
+        
+        multiPickerView.highlight(0, animated: false)
+    }
+    @objc func selected(_ sender: UIMultiPicker) {
+        
+        Logs.show(message: "Selected Index: \(sender.selectedIndexes)")
+        
+        selectedReasonsAray = sender.selectedIndexes
+        Logs.show(message: "Selected REASONS: \(selectedReasonsAray)")
+        
+    }
     
     @objc func removeBtnPressed(sender:UIButton) {
         let index = sender.tag
-        ApiService.deleteSocialLink(val: socialAccModel[index].socialAccountId)
-        socialAccModel.remove(at: index)
-        socialLinksTV.reloadData()
+        if canEdit {
+            ApiService.deleteSocialLink(val: socialAccModel[index].socialAccountId)
+            socialAccModel.remove(at: index)
+            socialLinksTV.reloadData()
+        } else {
+            setupMultiPickerView()
+        }
+        
+    }
+   
+    //MARK: API METHODS
+
+    func getReasons() {
+        
+        
+        APIService
+            .singelton
+            .getReportReasons()
+            .subscribe({[weak self] model in
+                guard let self = self else {return}
+                switch model {
+                    case .next(let val):
+                        AppFunctions.setIsNotifCheck(value: false)
+                        if val.count > 0 {
+                            self.reasonsList = val
+                            self.reasonsListName = self.reasonsList.map({$0.reason})
+                        } else {
+                            self.hidePKHUD()
+                        }
+                    case .error(let error):
+                        print(error)
+                        self.hidePKHUD()
+                    case .completed:
+                        print("completed")
+                        self.hidePKHUD()
+                }
+            })
+            .disposed(by: dispose_Bag)
     }
     
     
@@ -62,19 +132,26 @@ extension SocialLinks_VC : UITableViewDelegate, UITableViewDataSource {
         let cell : SocialAccTVCell = tableView.dequeueReusableCell(withIdentifier: "SocialAccTVCell", for: indexPath) as! SocialAccTVCell
         
         cell.socialLbl.text = socialAccModel[indexPath.row].linkTitle.capitalized
-        //cell.socialImgView.image = UIImage(named: linkType)
+        
         if socialAccModel[indexPath.row].linkImage != "" {
             let imageUrl = URL(string: socialAccModel[indexPath.row].linkImage)
             cell.socialImgView.sd_setImage(with: imageUrl , placeholderImage: UIImage()) { (image, error, imageCacheType, url) in }
         }
         cell.socialLbl.isUserInteractionEnabled = false
         
+        if isFromOther {
+            cell.removeBtn.isHidden = false
+        }
+        cell.removeBtn.tintColor = UIColor(named: "warning")
+        cell.removeBtn.backgroundColor = UIColor.clear
+        cell.removeBtn.setImage(UIImage(systemName: "exclamationmark.shield.fill"), for: .normal)
+        cell.removeBtn.tag = indexPath.row
+        cell.removeBtn.addTarget(self, action: #selector(removeBtnPressed(sender:)), for: .touchUpInside)
+        
         if canEdit {
             cell.removeBtn.isHidden = false
-            cell.removeBtn.addTarget(self, action: #selector(removeBtnPressed(sender:)), for: .touchUpInside)
             cell.removeBtn.setImage(UIImage(systemName: "xmark"), for: .normal)
             cell.removeBtn.cornerRadius = 10
-            cell.removeBtn.tag = indexPath.row
         }
         
         return cell
